tools/user_management/requirements.txt
auth0-python
boto3
